pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "global-logistics-resolver"
version = "1.0.0"
description = "High-stakes logistics simulator (OpenEnv) for SLA prioritization and routing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.0,<1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "requests>=2.31.0",
    "openai>=1.12.0",
]

[project.optional-dependencies]
plots = ["matplotlib>=3.8"]

[project.scripts]
server = "server.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["server*"]

[tool.setuptools]
py-modules = ["env", "models", "tasks"]
